Fix for mac and win64 build errors and also removing unnecessary code from ThreadsPosix.cpp

+++ VMPI/ThreadsWin.cpp
@@ -46,7 +46,7 @@ bool VMPI_tlsCreate(uintptr_t* tlsId)
 
 void VMPI_tlsDestroy(uintptr_t tlsId)
 {
-	TlsFree(tlsId);
+	TlsFree((DWORD)tlsId);
 }
 
 bool VMPI_tlsSetValue(uintptr_t tlsId, void* value)
